requirements.txt
autopep8==1.6.0
azure-ai-textanalytics==5.1.0
azure-common==1.1.28
azure-core==1.23.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
distlib==0.3.4
filelock==3.6.0
Flask==2.1.1
Flask-Cors==3.0.10
Flask-JSON==0.3.4
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
msrest==0.6.21
oauthlib==3.2.0
pipenv==2022.4.8
platformdirs==2.5.1
pycodestyle==2.8.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.9
virtualenv==20.14.0
virtualenv-clone==0.5.7
Werkzeug==2.1.1
zipp==3.8.0
